Project Creation
Project created in Unity.
First "Ring creation models" made in Blender, then imported to Unity as .fbx file.

Started "Level-building" process. Platforms spawn in a ring around the center of the level. Took a lot of time to troubleshoot some of the problems I was having with it, but it seems good enough for now.
For now it's just spawning platforms randomly from an array. Next, I need to give it some rules for spawning/not spawning certain platforms in certain places.