Basic Implementations
Set up the initial scene (maw, base ring, and player placeholders).
The LevelBuilder spawns a few sets of ringed platforms above the player's starting position.
The player can traverse around the rings, moving and jumping. Their rotation is basically working (which took FOR-FUCKING-EVER), but needs to be smoothed.
The camera tracks the player as they make their way around the rings. This also took a lot of time to figure out, but was actually easy to fix once I realized what the problem was.

Spent about the past 4 hours just trying to get the camera, and the player's rotation correct. Tried a LOT of different things that didn't work. Very annoying, but I'm glad these things are (basically) implemented, as they are a pretty damn big part of the game.